moved nextdate to enterprise
